geo-spark-ml/pyproject.toml
[project]
name = "geo-spark-ml"
version = "0.1.0"
description = "Geospatial clustering demo using PySpark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyspark==3.5.0",
    "pyyaml==6.0.1",
    "numpy==1.26.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geo_clustering"]
